refactor: lease-process as standalone module, agreement merged into market module
