+++ src/activity/work/work.ts
@@ -19,9 +19,9 @@ import { NetworkNode } from "../../network";
 import { RemoteProcess } from "./process";
 import { GolemWorkError, WorkErrorCode } from "./error";
 import { GolemConfigError, GolemTimeoutError } from "../../shared/error/golem-error";
-import { ProviderInfo } from "../../agreement";
+import { ProviderInfo } from "../../market/agreement";
 import { TcpProxy } from "../../network/tcpProxy";
-import { AgreementDTO } from "../../agreement/service";
+import { AgreementDTO } from "../../market/agreement/service";
 import { ActivityApi } from "ya-ts-client";
 import { YagnaExeScriptObserver } from "../../shared/yagna";
 import { ExeScriptExecutor } from "../exe-script-executor";
