+++ src/lease-process/lease-process-pool.ts
@@ -1,9 +1,9 @@
-import type { IAgreementApi, LegacyAgreementServiceOptions } from "./agreement";
+import type { IAgreementApi, LegacyAgreementServiceOptions } from "../market/agreement/agreement";
 import type { Logger } from "../shared/utils";
 import { defaultLogger } from "../shared/utils";
 import type { DraftOfferProposalPool, MarketModule } from "../market";
 import { GolemMarketError, MarketErrorCode } from "../market";
-import type { AgreementDTO } from "./service";
+import type { AgreementDTO } from "../market/agreement/service";
 import { EventEmitter } from "eventemitter3";
 import type { RequireAtLeastOne } from "../shared/utils/types";
 import type { Allocation, IPaymentApi } from "../payment";
