+++ src/market/agreement/strategy.ts
