+++ src/market/proposals_batch.test.ts
@@ -1,7 +1,7 @@
 import { ProposalsBatch } from "./proposals_batch";
 import { mock, instance, when } from "@johanblumenberg/ts-mockito";
 import { OfferProposal, ProposalProperties } from "./offer-proposal";
-import { ProviderInfo } from "../agreement";
+import { ProviderInfo } from "./agreement";
 
 const mockedProviderInfo: ProviderInfo = {
   id: "provider-id-1",
