+++ src/shared/yagna/adapters/activity-api-adapter.ts
@@ -1,4 +1,4 @@
-import { Agreement } from "../../../agreement";
+import { Agreement } from "../../../market/agreement";
 import { ActivityApi } from "ya-ts-client";
 import { Activity, ActivityStateEnum, GolemWorkError, IActivityApi, WorkErrorCode } from "../../../activity";
 import { IActivityRepository } from "../../../activity/activity";
